Delete scheduled calls when building is mark as not for sale.

+++ src/scheduled-events/event-listeners.js
@@ -1,5 +1,14 @@
+import { BUILDING_NEGOTIATION_STATUS_CHANGED } from '../building/service/UpdateBuildingNegotiationStatusService'
+
 export function setupEventListeners (eventBus, { scheduledCallRepository }) {
   eventBus.on('meeting.created', ({ buildingId }) => {
+    return scheduledCallRepository.removeScheduledCallsForBuilding(buildingId)
+  })
+
+  eventBus.on(BUILDING_NEGOTIATION_STATUS_CHANGED, ({ buildingId, negotiationStatus }) => {
+    if (negotiationStatus !== 'NO VENDE') {
+      return Promise.resolve()
+    }
     return scheduledCallRepository.removeScheduledCallsForBuilding(buildingId)
   })
 }
